move VariantSourceEntity to VariationCommons
